requirements.txt
APScheduler==3.9.1
async-timeout==4.0.2
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colorhash==1.1.0
commonmark==0.9.1
configparser==5.3.0
coverage==5.5
cryptography==38.0.1
dataclasses-json==0.5.7
Deprecated==1.2.13
deprecation==2.1.0
eyed3==0.9.6
filetype==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MonitoringDashboard==3.1.1
flask-prometheus-metrics==1.0.0
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mutagen==1.45.1
mypy-extensions==0.4.3
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==1.0.14
psutil==5.9.1
py-deezer==1.1.4.post1
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.13.0
PyInquirer==1.0.3
pyparsing==3.0.9
python-dotenv==0.20.0
python-youtube==0.8.2
pytz==2022.2.1
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
scipy==1.9.1
six==1.16.0
spotipy==2.20.0
SQLAlchemy==1.4.40
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.8.0
tzlocal==2.0.0
urllib3==1.26.11
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
youtube-dl==2020.12.22
youtube-search==2.1.1
yt-dlp==2022.8.14
